Split separate components for Mobile/Tablet and Desktop
