+++ packages/frontend/src/components/NavBar/DesktopDrawer.tsx
@@ -1,0 +1,96 @@
+import { ListItem, ListItemButton, ListItemIcon, Toolbar, Typography } from '@mui/material';
+import MenuIcon from '@mui/icons-material/Menu';
+import ArrowBackIcon from '@mui/icons-material/ArrowBack';
+import styled from '@emotion/styled';
+import MuiDrawer, { DrawerProps } from '@mui/material/Drawer';
+import { Theme, CSSObject } from '@mui/material/styles';
+
+import { Colors } from '../../constants/Colors';
+import RataExtLogo from '../../assets/images/Logo_noText.png';
+// import { DrawerWrapper } from './index.styles';
+import { MenuList } from './MenuList';
+import { drawerWidth } from '../../constants/Viewports';
+
+type DesktopDrawerProps = {
+  open: boolean;
+  toggleDrawer: React.MouseEventHandler<HTMLElement>;
+};
+
+export const DesktopDrawer = ({ open, toggleDrawer }: DesktopDrawerProps) => {
+  return (
+    <DesktopDrawerWrapper variant="permanent" anchor="left" open={open}>
+      <Toolbar>
+        <Typography component="img" src={RataExtLogo} alt="Logo" sx={{ width: '65px', height: '65px' }} />
+        <Typography sx={{ fontSize: '18px', opacity: open ? 1 : 0 }}>Ratatiedon extranet</Typography>
+      </Toolbar>
+      <ListItem key={open ? 'Close drawer' : 'Open drawer'} disablePadding onClick={toggleDrawer}>
+        <ListItemButton area-label={open ? 'close drawer' : 'open drawer'}>
+          <ListItemIcon>{open ? <ArrowBackIcon color="primary" /> : <MenuIcon color="primary" />}</ListItemIcon>
+        </ListItemButton>
+      </ListItem>
+      <MenuList open={open} />
+    </DesktopDrawerWrapper>
+  );
+};
+
+const openedMixin = (theme: Theme): CSSObject => ({
+  width: drawerWidth,
+  transition: theme.transitions.create('width', {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.enteringScreen,
+  }),
+  overflowX: 'hidden',
+});
+
+const closedMixin = (theme: Theme): CSSObject => ({
+  transition: theme.transitions.create('width', {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.leavingScreen,
+  }),
+  overflowX: 'hidden',
+  width: `calc(${theme.spacing(7)} + 1px)`,
+  [theme.breakpoints.up('tablet')]: {
+    width: `calc(${theme.spacing(8)} + 1px)`,
+  },
+});
+
+interface DrawerWrapperProps extends DrawerProps {
+  open?: boolean;
+}
+export const DesktopDrawerWrapper = styled(MuiDrawer)<DrawerWrapperProps>(({ theme, open }) => {
+  return {
+    [theme.breakpoints.down('desktop')]: {
+      display: 'none',
+    },
+    [theme.breakpoints.up('desktop')]: {
+      '& .MuiPaper-root': {
+        borderColor: Colors.lightblue,
+        borderWidth: '3px',
+      },
+      flexShrink: 0,
+      whiteSpace: 'nowrap',
+      boxSizing: 'border-box',
+      ...((open && {
+        ...openedMixin(theme),
+        '& .MuiDrawer-paper': openedMixin(theme),
+      }) as any),
+      ...((!open && {
+        ...closedMixin(theme),
+        '& .MuiDrawer-paper': closedMixin(theme),
+      }) as any),
+
+      // Logout is the last list item
+      '& li:last-child': {
+        position: 'fixed',
+        bottom: '16px',
+        width: open ? `${drawerWidth}px` : `calc(${theme.spacing(8)} + 1px)`,
+        '& .MuiListItemIcon-root': {
+          color: Colors.darkblue,
+        },
+        '& .MuiListItemText-root': {
+          color: Colors.darkblue,
+        },
+      },
+    },
+  };
+});
