--- packages/frontend/src/components/NavBar/index.styles.ts
@@ -1,97 +1,0 @@
-import styled from '@emotion/styled';
-import MuiAppBar, { AppBarProps } from '@mui/material/AppBar';
-import { Theme, CSSObject } from '@mui/material/styles';
-import MuiDrawer, { DrawerProps } from '@mui/material/Drawer';
-
-import { Colors } from '../../constants/Colors';
-
-interface AppBarWrapperProps extends AppBarProps {
-  open?: boolean;
-}
-
-const drawerWidth = 306;
-
-export const MiniAppBarWrapper = styled(MuiAppBar, {
-  shouldForwardProp: (prop) => prop !== 'open',
-})<AppBarWrapperProps>(({ theme, open }) => {
-  return {
-    zIndex: theme.zIndex.drawer + 1,
-    backgroundColor: Colors.white,
-    [theme.breakpoints.down('tablet')]: {
-      boxShadow: `0px 3px ${Colors.lightblue}`,
-    },
-    [theme.breakpoints.up('tablet')]: {
-      boxShadow: `${open ? `${drawerWidth}px` : '0px'} 3px ${Colors.lightblue}`,
-    },
-    [theme.breakpoints.up('desktop')]: {
-      display: 'none',
-    },
-  };
-});
-
-const openedMixin = (theme: Theme): CSSObject => ({
-  width: drawerWidth,
-  transition: theme.transitions.create('width', {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.enteringScreen,
-  }),
-  overflowX: 'hidden',
-});
-
-const closedMixin = (theme: Theme): CSSObject => ({
-  transition: theme.transitions.create('width', {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  overflowX: 'hidden',
-  width: `calc(${theme.spacing(7)} + 1px)`,
-  [theme.breakpoints.up('tablet')]: {
-    width: `calc(${theme.spacing(8)} + 1px)`,
-  },
-});
-
-interface DrawerWrapperProps extends DrawerProps {
-  open?: boolean;
-}
-
-export const DrawerWrapper = styled(MuiDrawer)<DrawerWrapperProps>(({ theme, open }) => {
-  return {
-    [theme.breakpoints.down('tablet')]: {
-      '& .MuiPaper-root': {
-        width: '100%',
-      },
-    },
-    [theme.breakpoints.up('tablet')]: {
-      '& .MuiPaper-root': {
-        width: `${drawerWidth}px`,
-        borderColor: Colors.lightblue,
-        borderWidth: '3px',
-      },
-    },
-    [theme.breakpoints.up('desktop')]: {
-      flexShrink: 0,
-      whiteSpace: 'nowrap',
-      boxSizing: 'border-box',
-      ...((open && {
-        ...openedMixin(theme),
-        '& .MuiDrawer-paper': openedMixin(theme),
-      }) as any),
-      ...((!open && {
-        ...closedMixin(theme),
-        '& .MuiDrawer-paper': closedMixin(theme),
-      }) as any),
-    },
-    // Logout is the last list item
-    '& li:last-child': {
-      position: 'fixed',
-      bottom: '16px',
-      width: open ? `${drawerWidth}px` : `calc(${theme.spacing(8)} + 1px)`,
-      '& .MuiListItemIcon-root': {
-        color: Colors.darkblue,
-      },
-      '& .MuiListItemText-root': {
-        color: Colors.darkblue,
-      },
-    },
-  };
-});
